requirements.txt
# Core AI/ML libraries
tensorflow>=2.12.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0

# Image processing
opencv-python>=4.7.0
Pillow>=9.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web app
streamlit>=1.22.0

# Dataset download
kaggle>=1.5.13

# Utilities
tqdm>=4.65.0

streamlit
protobuf==3.20.3
requests
fastapi
uvicorn
python-multipart